app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
aiosqlite>=0.20.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
playwright>=1.49.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
itsdangerous>=2.2.0
Pillow>=10.0.0
qrcode>=7.4.0
# ── Ativação por token (criptografia AES-256-GCM) ────────────────────────────
cryptography>=42.0.0
# ── Kiosk window sem barra de endereços (WebView2) ──────────────────────────
# Instalado apenas no launcher_gui.exe — não necessário para o serviço Windows
pywebview>=5.0.0
# ── Worker standalone (ZapDin-Worker.exe) ────────────────────────────────────
# httpx já incluído acima
